Undid accidental commit for some files.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@33653 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/threading/thread.h
@@ -75,7 +75,7 @@ namespace Threading
   };
 
   /// OS specific thread identifier
-  typedef unsigned long ThreadID;
+  typedef unsigned int ThreadID;
 
 }
 }
